changing int32_t to int64_t to handle leage sizes

+++ exa-defs.h
@@ -78,7 +78,7 @@
 #define FILE_NAME_LEN 1024
 #define TOLTEST 1e-9
 
-typedef int32_t emInt;
+typedef int64_t emInt;
 #define EMINT_MAX UINT_MAX
 
 #if (HAVE_CGNS == 0)
